+++ src/components/UserInfo.js
@@ -1,9 +1,9 @@
-import { api } from "./cards";
-
 export class UserInfo {
   constructor(nameSelector, aboutSelector, getInfo, setInfo) {
     this.nameSelector = nameSelector;
     this.aboutSelector = aboutSelector;
+    this._id;
+    this.avatar;
     this._getInfo = getInfo;
     this._setInfo = setInfo;
   }
@@ -17,27 +17,3 @@ export class UserInfo {
     this._setInfo(name, about, this.nameSelector, this.aboutSelector);
   }
 }
-
-const userInfo = new UserInfo('.profile__name', '.profile__bio', () => {
-  return api.getUserInformation()
-    .then((res) => {
-      const user = {};
-      user.name = res["name"];
-      user.about = res["name"];
-      return user;
-    })
-    .catch(err => {
-      console.log(`Ошибка: ${err}`);
-    });
-}, (name, about, nameSelector, aboutSelector) => {
-  api.editProfile(name, about)
-    .then((res) => {
-      document.querySelector(nameSelector).textContent = res["name"];
-      document.querySelector(aboutSelector).textContent = res["about"];
-    })
-    .catch(err => {
-      console.log(`Ошибка: ${err}`);
-    });
-})
-
-userInfo.getUserInfo().then((res) => console.log(res));
