+++ src/utils/constants.js
@@ -1,0 +1,8 @@
+const validationConfig = {
+  formSelector: '.popup__input-container',
+  inputSelector: '.popup__input-item',
+  submitButtonSelector: '.popup__save-button',
+  inactiveButtonClass: 'popup__save-button_inactive',
+  inputErrorClass: 'popup__input-item_type_error',
+  errorClass: 'popup__input-item-error_active'
+};
